+++ inst/templates/libraries.R
@@ -1,0 +1,35 @@
+# Twitter
+library("rtweet")
+
+# Data manipulation
+library("dplyr")
+library("tidyr")
+library("lubridate")
+library("forcats")
+library("purrr")
+library("stringr")
+
+# Networks
+library("igraph")
+
+# Text analysis
+library("tidytext")
+library("topicmodels")
+
+# Visualisation
+library("ggplot2")
+library("ggrepel")
+library("ggraph")
+library("wordcloud")
+library("viridis")
+
+# Images
+library("webshot")
+library("magick")
+
+# Presentation
+library("knitr")
+
+# File paths
+library("here")
+library("fs")